Add render optimizations
