+++ project/src/pages/film/film.tsx
@@ -1,36 +1,20 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { generatePath, Link, useParams } from 'react-router-dom';
 
 import { Logo } from '../../components/logo/logo';
 import { AppRoute } from '../../constants/routs';
 import { SimilarFilms } from '../../components/similar-films/similar-films';
 import { IconAdd, IconPlayS } from '../../components/icon';
-import { Tabs } from '../../components/tabs';
-import { FilmDetails } from '../../components/film-details/film-details';
-import { FilmReviews } from '../../components/film-reviews/film-reviews';
-import { FilmOverview } from '../../components/film-overview/film-overview';
 import { UserBlock } from '../../components/user-block/user-block';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { fetchFilm, fetchSimilarFilms } from '../../store/api-actions';
 import { LoadingScreen } from '../../components/loading-screen/loading-screen';
 import { AuthorizationStatus } from '../../constants/auth';
-
-enum FilmTab {
-  Overview = 'Overview',
-  Details = 'Details',
-  Reviews = 'Reviews',
-}
-
-const tabs = [
-  { title: FilmTab.Overview },
-  { title: FilmTab.Details },
-  { title: FilmTab.Reviews },
-];
+import { FilmDescription } from '../../components/film-description/film-description';
 
 function Film(): JSX.Element {
   const dispatch = useAppDispatch();
   const { id: filmId } = useParams();
-  const [active, setActive] = useState<string>(FilmTab.Overview);
 
   const { film, similarFilms, authorizationStatus } = useAppSelector((state) => state);
 
@@ -107,13 +91,7 @@ function Film(): JSX.Element {
               />
             </div>
 
-            <div className="film-card__desc">
-              <Tabs tabs={tabs} active={active} onChange={setActive} />
-
-              {active === FilmTab.Overview && <FilmOverview film={film.data} />}
-              {active === FilmTab.Details && <FilmDetails film={film.data} />}
-              {active === FilmTab.Reviews && <FilmReviews />}
-            </div>
+            <FilmDescription film={film.data} />
           </div>
         </div>
       </section>
